OpTestSSH: Make an SSH helper common to all platforms
Several classes separately implement pxssh helpers, and OpenBMC has its
own ssh-pexpect implementation as well. Instead of trying to maintain
all these different classes create a single OpTestSSH class. This
essentially implements the OpenBMC SSH approach which means we drop the
pxssh dependency and automatically inherit the useful OPexpect class.
